Environment for GitHub Web Hooks processing
This commit implements an environment for GitHub Web Hooks processing.

Details of how everything works are in the README.

I attempted to import existing state from the moz-devservices
account. However, not all items could be imported. I'm tempted
to delete everything from the AWS account and use Terraform
to define everything from a clean slate to make sure it works.

We should probably coordinate on this pull request landing, as I'm
not sure if any of this works. `terraform plan` seems reasonable,
however.