Convert IAM policies to aws_iam_policy_document data
Before, we used inline JSON to define the IAM policies. Today I
learned about the "aws_iam_policy_document" Terraform data type,
which provides a more structured way of defining IAM policies.

This commit switches to the data type for declaring IAM policies.
A consequence of this change is we can add trailing commas and
inline comments documenting the policy in more detail.

As part of this, I also tweaked the CloudWatch actions to allow
log group and stream creation.