hgmo: s3 bucket for holding events (bug 1316952)
We want hg.mo to write a stream of log events to S3 for data
retention and for reference by other services. Some events are on
the larger side and exceed size limits in SNS and Kinesis Firehose.
So, we'll write all events to S3 and then reference the object URL
in SNS/Kinesis/etc events if the event is too large to fit in a payload.

This commit introduces an S3 bucket and policy to hold hg.mo
events. The hgnotifier user can write to the bucket. Anybody can
read from the bucket (because all data is public and already posted
to Pulse).