vcssync: new environment to manage version control syncing
We need to deploy a dedicated instance to perform version control
syncing operations in support of Stylo and other random efforts (like
a Firefox Mercurial repository with full CVS history).

While we have an existing "vcs sync" EC2 instance, this is running
code that is completely independent from what I'll be deploying. I'd
like to keep the instances separated.

In this commit, we establish the scaffolding for the new "vcssync"
environment.