vcssync: configuration for running Servo VCS sync
We need an EC2 instance for running synchronization processes.
This commit introduces that instance.

We also need an S3 bucket to hold state. This commit introduces that
as well.

IAM policies are hard. We create a new IAM role for the EC2 instance
that grants it access to the S3 bucket for SSH keys and our S3 state
bucket. For the SSH key bucket, we needed to add this role to the
bucket policy in the base environment.

Rounding it out, we add a Route53 hostname to make connecting to
the instance easier.