Bug 1374380 - Temporary hostname for build metrics ingestion
This is meant to be temporary (weeks to months). Name was chosen
somewhat arbitrarily. The service this is related to doesn't
have a presence in our AWS account. So there was no obvious file
to define the record in. Since it is meant to be temporary,
I just put it in the base "profile."