Bug 1381391 - SNS topic for Servo VCS sync errors
We need an SNS topic for holding errors related to Servo VCS sync.
We plan to use the SNS topic to alert humans when errors occur.

This commit defines a new SNS topic. It is writable by the VCS Sync
EC2 instance. The world can subscribe to it (there should be no
private or sensitive data leaked in the error messages).