global: migrate to backends
Terraform 0.9 introduced backends. Backends are apparently the
new way to do remote state.

This commit upgrades everything to use backends.

Essentially, this meant adding a backend {} in resources.tf and
updating init.sh to use `terraform init` instead of `terraform remote`.
Since most backend configs are constant, init.sh is simplified to
only specify the per-environment S3 key to use.

On my machine, `runtests.sh` was failing on 0.10. This appears to
be a change in behavior in either 0.9 or 0.10 with regards to
validating directories that don't have a backend configured.

The root directory has no explicit state/resources. So we exclude
it from testing.