hgmo: update CloudFront certificate
We uploaded a new certificate for the CloudFront distribution a few
weeks back. It looks like this was done manually because terraform
is still referencing the old certificate ID.

This commit points terraform at the ARN of the new ACM-hosted
certificate. After this change, `terraform plan` for hgmo says no
changes are necessary, as expected.