vcssync: give wpt sync testing instance access to ssh keys S3 bucket
We need to give the instance access to the S3 bucket with SSH keys
in order for SSH key synchronization to work. We accomplish this
using an instance profile, just like for Servo VCS Sync.

We also needed to update the access policy on the S3 bucket to
allow the new role to read from the bucket. We should probably
consolidate this down to a single named policy that can be used
by multiple Terraform projects. But that's follow-up fodder.