Bug 1425593 - Create S3 bucket for Mercurial bundles in use2
We apparently have TaskCluster CI running in us-east-2. We should
be hosting Mercurial bundles from there so all tasks in that region
don't incur cross-region traffic.

As part of rolling this out, I discovered that we apparently have
a reliance on an ACL defined via the web console. The short version
is that "Everyone" can "list objects." This appears to effectively
be granting the "s3:ListBucket" permission. I added this permission
to all IAM policies.

There is a lot of copy pasta in these files. I tried to refactor
things to use "count" and variable expansion so we only need to
define each resource once. However, terraform doesn't seem to
allow the "provider" variable to be dynamic. I suspect we could
use templates or modules. But this is beyond my terraform knowledge.
So we incur the copy pasta.