Define an ACL policy for S3 buckets
An empty policy string appears to preserve whatever changes are made
in the AWS web console.

When deploying the hg bundles bucket in us-east-2, I uncovered that
existing buckets appeared to have manual ACL modifications made
in the AWS web console.

We have since updated the IAM policies for the buckets to confer
the appropriate permissiosn to provide read-only access to all
users.

In this commit, we set the ACL policy of the bucket to "private."
This isn't technically the "correct" ACL since the bucket is public.
But "private" is the default ACL and all permissions should be
granted via IAM policy, not ACL, as ACLs are kinda deprecated. So
"private" (no access by default) feels like the proper policy.