treeherder: Allow recreating dev RDS from latest prod snapshot
Uses the new `aws_db_snapshot` data source added in Terraform v0.9.6:
https://www.terraform.io/docs/providers/aws/d/db_snapshot.html
https://github.com/hashicorp/terraform/commit/2276e981ee3955d0475c8a71f3634c784a7f710b

The lifecycle `ignore_changes` entry is required because the
`snapshot_identifer` property is marked as `ForceNew`:
https://github.com/terraform-providers/terraform-provider-aws/blob/v1.10.0/aws/resource_aws_db_instance.go#L285-L291

With this configuration the dev instance should remain untouched
unless a manual `terraform taint` is performed on the resource, at
which point it will be recreated using the latest prod snapshot at
the time.

Stage will be updated similarly, presuming no problems are seen on dev.

Note: terraform-provider-aws v1.3.0+ is recommended, to avoid issues
creating an instance when the latest snapshot creation is in progress:
https://github.com/terraform-providers/terraform-provider-aws/commit/0319feffc0d93c6f62cba2ae4b988e5bfd294c00

Fixes bug 1321294.