Bug 1453553 - Remove most Servo resources
These resources moved elsewhere. We can safely remove all
references to Servo VCS sync from our AWS account.

We'll want to apply this change first then delete the provider
later, as the provider is effectively empty after this. If we
deleted the provider outright, we wouldn't be able to run
terraform to delete resources!