Add gps2 user
IAM users can only have up to 2 access keys. I like to use separate
keys per physical device so I can more easily audit for key usage
and so if/when a key is revoked, I'm not locked out of access on all
my devices.

So let's create a gps2 user so it can hold an additional 2 access keys.